refactor: replace AppConfig with ExtHostAppConfig
