+++ packages/extension/src/hosted/api/vscode/debug/ext.host.debug.ts
@@ -7,6 +7,7 @@ import {
   IDebuggerContribution,
 } from '@opensumi/ide-debug';
 
+import { CustomChildProcessModule } from '../../../../common/ext.process';
 import {
   IExtHostCommands,
   IExtHostDebugService,
@@ -25,7 +26,6 @@ import {
   Uri,
 } from '../../../../common/vscode/ext-types';
 import { Breakpoint } from '../../../../common/vscode/models';
-import { CustomChildProcessModule } from '../../../ext.process-base';
 
 import { IDebugConfigurationProvider } from './common';
 import { resolveDebugAdapterExecutable } from './extension-debug-adapter-excutable-resolver';
