+++ packages/logs-core/src/node/log-manager.ts
@@ -1,6 +1,6 @@
 import path from 'path';
 
-import { Autowired, ConstructorOf, Injectable } from '@opensumi/di';
+import { ConstructorOf, Injectable, Optional } from '@opensumi/di';
 import { Emitter } from '@opensumi/ide-core-common';
 import { AppConfig } from '@opensumi/ide-core-node/lib/types';
 
@@ -19,9 +19,6 @@ import { cleanAllLogs, cleanExpiredLogs, cleanOldLogs, getLogFolder, getLogZipAr
 
 @Injectable()
 export class LogServiceManager implements ILogServiceManager {
-  @Autowired(AppConfig)
-  private appConfig: AppConfig;
-
   protected readonly logLevelChangeEmitter = new Emitter<LogLevel>();
   private globalLogLevel: LogLevel;
   private logMap = new Map<SupportLogNamespace, ILogService>();
@@ -29,7 +26,7 @@ export class LogServiceManager implements ILogServiceManager {
   private logFolderPath: string;
   private LogServiceClass: ConstructorOf<ILogService>;
 
-  constructor() {
+  constructor(@Optional(AppConfig) private appConfig: AppConfig) {
     this.init({
       logDir: this.appConfig.logDir,
       logLevel: this.appConfig.logLevel,
